+++ packages/workeros-programs/src/index.js
@@ -102,6 +102,7 @@ export const programs = [
  * @type {{ path: string, source: () => Promise<string> }[]}
  */
 export const libraries = [
+  { path: "/lib/workeros-cli/args.js", source: () => fetchText("./cli/args.js") },
   { path: "/lib/workeros-node/require-runtime.js", source: () => fetchText("./node/require-runtime.js") },
   { path: "/lib/workeros-node/fs.js", source: () => fetchText("./node/fs.js") },
   { path: "/lib/workeros-node/path.js", source: () => fetchText("./node/path.js") },